281: move applicable constants into subfolders
