+++ src/control/physics.js
@@ -1,4 +1,5 @@
-import { BRP, ACCEL, MAX_SPEED, DAMPING_FACTOR } from '../constants';
+import { BRP } from '../constants';
+import { ACCEL, MAX_SPEED, DAMPING_FACTOR } from './constants';
 import { assert } from '../utils/asserts';
 import { getMe } from '../look/gameState';
 import { FSM } from '../think/fsm';
