+++ src/draw/constants.js
@@ -1,0 +1,16 @@
+export const ALLY_PATH_COLOR = 0x00ff00; // Green
+export const ENEMY_PATH_COLOR = 0x993333; // Dark red
+export const NT_ALPHA = 0.4;
+export const NT_COLOR = 0xff8c00; // Orange
+export const KEY_COLOR = 0x753daf; // Purple
+export const KEY_ON_ALPHA = 0.75;
+export const KEY_OFF_ALPHA = 0.2;
+export const NAV_MESH_EDGE_COLOR = 0xff0090; // Pink
+export const NAV_MESH_FIXED_EDGE_COLOR = 0x42aaf4; // Bright blue
+export const NAV_MESH_VERTEX_COLOR = 0xa5005d; // Darker Pink
+export const NAV_MESH_ALPHA = 0.2;
+export const NAV_MESH_FIXED_EDGE_ALPHA = 1;
+export const NAV_MESH_THICKNESS = 4; // Line thickness in pixels
+export const TRIANGULATION_EDGE_COLOR = 0x006600; // Turquoise
+export const TRIANGULATION_ALPHA = 0.2;
+export const TRIANGULATION_THICKNESS = 4;
