+++ src/look/constants.js
@@ -1,0 +1,4 @@
+export const TEAMS = {
+  RED: 1,
+  BLUE: 2,
+};
